+++ spring-cloud-starter-tencent-polaris-config/src/test/java/com/tencent/cloud/polaris/config/adapter/PolarisConfigFilePullerTest.java
@@ -27,14 +27,14 @@ import com.tencent.cloud.polaris.config.config.ConfigFileGroup;
 import com.tencent.cloud.polaris.context.config.PolarisContextProperties;
 import com.tencent.polaris.configuration.api.core.ConfigFileService;
 import com.tencent.polaris.configuration.api.core.ConfigKVFile;
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import org.springframework.core.env.CompositePropertySource;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
 /**
@@ -42,7 +42,7 @@ import static org.mockito.Mockito.when;
  *
  * @author wlx
  */
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class PolarisConfigFilePullerTest {
 
 	private final String testNamespace = "testNamespace";
@@ -81,9 +81,9 @@ public class PolarisConfigFilePullerTest {
 
 		puller.initInternalConfigFiles(compositePropertySource, new String[] {}, new String[] {}, testServiceName);
 
-		Assert.assertEquals("v1", compositePropertySource.getProperty("k1"));
-		Assert.assertEquals("v2", compositePropertySource.getProperty("k2"));
-		Assert.assertEquals("v3", compositePropertySource.getProperty("k3"));
+		assertThat(compositePropertySource.getProperty("k1")).isEqualTo("v1");
+		assertThat(compositePropertySource.getProperty("k2")).isEqualTo("v2");
+		assertThat(compositePropertySource.getProperty("k3")).isEqualTo("v3");
 	}
 
 	@Test
@@ -124,9 +124,9 @@ public class PolarisConfigFilePullerTest {
 		CompositePropertySource compositePropertySource = new CompositePropertySource(polarisConfigPropertySourceName);
 		puller.initInternalConfigFiles(compositePropertySource, activeProfiles, new String[] {}, testServiceName);
 
-		Assert.assertEquals("v11", compositePropertySource.getProperty("k1"));
-		Assert.assertEquals("v2", compositePropertySource.getProperty("k2"));
-		Assert.assertEquals("v3", compositePropertySource.getProperty("k3"));
+		assertThat(compositePropertySource.getProperty("k1")).isEqualTo("v11");
+		assertThat(compositePropertySource.getProperty("k2")).isEqualTo("v2");
+		assertThat(compositePropertySource.getProperty("k3")).isEqualTo("v3");
 	}
 
 	@Test
@@ -162,8 +162,8 @@ public class PolarisConfigFilePullerTest {
 		CompositePropertySource compositePropertySource = new CompositePropertySource(polarisConfigPropertySourceName);
 		puller.initCustomPolarisConfigFiles(compositePropertySource, customFiles);
 
-		Assert.assertEquals("v1", compositePropertySource.getProperty("k1"));
-		Assert.assertEquals("v2", compositePropertySource.getProperty("k2"));
-		Assert.assertEquals("v3", compositePropertySource.getProperty("k3"));
+		assertThat(compositePropertySource.getProperty("k1")).isEqualTo("v1");
+		assertThat(compositePropertySource.getProperty("k2")).isEqualTo("v2");
+		assertThat(compositePropertySource.getProperty("k3")).isEqualTo("v3");
 	}
 }